Output binary PD annotations in binary mode on Windows
On Windows, stdout defaults to text mode, which attempts to perform
newline conversion such that if it sees a bare line feed (hex 0A),
it will insert a carriage return (hex 0D) before the line feed.

For actual binary output, this is likely to garble the output.

When the -B/ --protocol-decoder-binary option is active, change
the mode of stdout to _O_BINARY to disable this newline conversion
to ensure that the protocol decoder binary output is faithfully
delivered to stdout without modifications.